Move a modify_indexes::Db::update_items_count into a db actor
This is a part of #3.

The change moves update_items_count method from modify_indexes into a db
actor. Next paches will move other methods into a db actor.
