+++ src/modify_indexes.rs
@@ -4,7 +4,6 @@
  */
 
 use crate::IndexId;
-use crate::IndexItemsCount;
 use crate::db;
 use anyhow::Context;
 use scylla::client::session::Session;
@@ -15,29 +14,14 @@ use tokio::sync::mpsc::Sender;
 use tracing::warn;
 
 pub(crate) enum ModifyIndexes {
-    UpdateItemsCount {
-        id: IndexId,
-        items_count: IndexItemsCount,
-    },
-    Del {
-        id: IndexId,
-    },
+    Del { id: IndexId },
 }
 
 pub(crate) trait ModifyIndexesExt {
-    async fn update_items_count(&self, id: IndexId, items_count: IndexItemsCount);
     async fn del(&self, id: IndexId);
 }
 
 impl ModifyIndexesExt for Sender<ModifyIndexes> {
-    async fn update_items_count(&self, id: IndexId, items_count: IndexItemsCount) {
-        self.send(ModifyIndexes::UpdateItemsCount { id, items_count })
-            .await
-            .unwrap_or_else(|err| {
-                warn!("ModifyIndexesExt::update_items_count: unable to send request: {err}")
-            });
-    }
-
     async fn del(&self, id: IndexId) {
         self.send(ModifyIndexes::Del { id })
             .await
@@ -58,12 +42,6 @@ pub(crate) async fn new(
     tokio::spawn(async move {
         while let Some(msg) = rx.recv().await {
             match msg {
-                ModifyIndexes::UpdateItemsCount { id, items_count } => db
-                    .update_items_count(id, items_count)
-                    .await
-                    .unwrap_or_else(|err| {
-                        warn!("modify_indexes: unable to update items count for index in db: {err}")
-                    }),
                 ModifyIndexes::Del { id } => db.remove_index(id).await.unwrap_or_else(|err| {
                     warn!("modify_indexes: unable to remove index from db: {err}")
                 }),
@@ -75,39 +53,18 @@ pub(crate) async fn new(
 
 struct Db {
     session: Arc<Session>,
-    st_update_items_count: PreparedStatement,
     st_remove_index: PreparedStatement,
 }
 
 impl Db {
     async fn new(session: Arc<Session>) -> anyhow::Result<Self> {
         Ok(Self {
-            st_update_items_count: session
-                .prepare(Self::UPDATE_ITEMS_COUNT)
-                .await
-                .context("UPDATE_ITEMS_COUNT")?,
             st_remove_index: session
                 .prepare(Self::REMOVE_INDEX)
                 .await
                 .context("REMOVE_INDEX")?,
             session,
         })
-    }
-
-    const UPDATE_ITEMS_COUNT: &str = "
-        UPDATE vector_benchmark.vector_indexes
-            SET indexed_elements_count = ?
-            WHERE id = ?
-        ";
-    async fn update_items_count(
-        &self,
-        id: IndexId,
-        items_count: IndexItemsCount,
-    ) -> anyhow::Result<()> {
-        self.session
-            .execute_unpaged(&self.st_update_items_count, (items_count, id))
-            .await?;
-        Ok(())
     }
 
     const REMOVE_INDEX: &str = "DELETE FROM vector_benchmark.vector_indexes WHERE id = ?";
